// src/app/jobseeker/dashboard/components/jobseeker-profile/components/profile-experience/models/workExperience.ts
export class workDetails1{

    companyName: string = " ";
    jobTitle: string = "";
    industryType: string = " ";
    location: string = " ";
    description: string = " ";
    period: string = " ";
    department: string = " ";
    currentlyWorking:boolean= false;
    startDate: string = " ";
    endDate: string = " ";
    currentSalary:number ;
    salaryType: string = "MONTHLY";
    empType : string = "FULLTIME";
    experience:number;
}